Potato Leaf Disease Classification
This is a Deep Learning Project which uses Convolution Neural Networks which identifies the affected leaves of the potato plant and makes classification based on the predictions. The libraries used in this project are Tensorflow, Keras and Matplotlib. 

The Dataset is from Kaggle -> https://www.kaggle.com/datasets/rizwan123456789/potato-disease-leaf-datasetpld